Move the interactor invoker output to inner classes rather than using the interactor output injector that is now broken. Remove unneeded classes
