+++ app/src/main/java/me/panavtec/cleancontacts/domain/invoker/InteractorOutputTask.java
@@ -3,7 +3,7 @@ package me.panavtec.cleancontacts.domain.invoker;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.FutureTask;
-import me.panavtec.presentation.common.outputs.InteractorOutput;
+import me.panavtec.cleancontacts.presentation.InteractorOutput;
 
 public class InteractorOutputTask<T, E extends Exception> extends FutureTask<T>
     implements PriorizableInteractor {
