+++ presentation/src/main/java/me/panavtec/cleancontacts/presentation/modules/detail/DetailPresenter.java
@@ -5,12 +5,8 @@ import me.panavtec.cleancontacts.domain.interactors.contacts.GetContactInteracto
 import me.panavtec.cleancontacts.domain.interactors.contacts.exceptions.ObtainContactException;
 import me.panavtec.cleancontacts.presentation.invoker.InteractorInvoker;
 import me.panavtec.cleancontacts.presentation.model.mapper.PresentationContactMapper;
-import me.panavtec.presentation.Presenter;
-import me.panavtec.presentation.common.outputs.InteractorOutput;
-import me.panavtec.presentation.common.outputs.InteractorOutputInjector;
-import me.panavtec.presentation.common.outputs.qualifiers.OnError;
-import me.panavtec.presentation.common.outputs.qualifiers.OnResult;
-import me.panavtec.presentation.common.outputs.qualifiers.Output;
+import me.panavtec.cleancontacts.presentation.Presenter;
+import me.panavtec.cleancontacts.presentation.InteractorOutput;
 import me.panavtec.threaddecoratedview.views.ThreadSpec;
 
 public class DetailPresenter extends Presenter<DetailView> {
@@ -19,8 +15,6 @@ public class DetailPresenter extends Presenter<DetailView> {
   private final InteractorInvoker interactorInvoker;
   private final GetContactInteractor getContactInteractor;
   private final PresentationContactMapper presentationContactMapper;
-  @Output
-  InteractorOutput<Contact, ObtainContactException> getContactOutput;
 
   public DetailPresenter(String contactMd5, InteractorInvoker interactorInvoker,
       GetContactInteractor getContactInteractor,
@@ -30,7 +24,6 @@ public class DetailPresenter extends Presenter<DetailView> {
     this.interactorInvoker = interactorInvoker;
     this.getContactInteractor = getContactInteractor;
     this.presentationContactMapper = presentationContactMapper;
-    InteractorOutputInjector.inject(this);
   }
 
   @Override public void onViewAttached() {
@@ -43,16 +36,18 @@ public class DetailPresenter extends Presenter<DetailView> {
 
   public void obtainContact() {
     getContactInteractor.setData(contactMd5);
-    interactorInvoker.execute(getContactInteractor, getContactOutput);
-  }
+    interactorInvoker.execute(getContactInteractor,
+        new InteractorOutput<Contact, ObtainContactException>() {
+          @Override public void onResult(Contact result) {
+            getView().showContactData(presentationContactMapper.modelToData(result));
+          }
 
-  @OnResult
-  void onContactInteractor(Contact data) {
-    getView().showContactData(presentationContactMapper.modelToData(data));
-  }
+          @Override public void onError(ObtainContactException exception) {
+            getView().showGetContactError();
+          }
 
-  @OnError
-  void onContactInteractorError(ObtainContactException e) {
-    getView().showGetContactError();
+          @Override public void onCancel() {
+          }
+        });
   }
 }
