--- presentation/src/main/java/me/panavtec/presentation/common/ThreadSpec.java
@@ -1,5 +1,0 @@
-package me.panavtec.presentation.common;
-
-public interface ThreadSpec {
-  public void execute(Runnable action);
-}
