--- presentation/src/main/java/me/panavtec/presentation/common/outputs/qualifiers/OnResult.java
@@ -1,10 +1,0 @@
-package me.panavtec.presentation.common.outputs.qualifiers;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD) @Retention(RetentionPolicy.SOURCE)
-public @interface OnResult {
-}
